% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.recombinations.per.base.R
\name{count.rec.events.per.base}
\alias{count.rec.events.per.base}
\title{Count number of overlapping recombination events at each genomic position}
\usage{
\method{count}{rec.events.per.base}(gubbins.gff.file, recom.input.type = "Gubbins")
}
\arguments{
\item{gubbins.gff.file}{Path to the input Gubbins GFF recombination file or data frame}

\item{recom.input.type}{Type of input recombination data, either "Gubbins" GFF or "BRATNextGen" tabular data.}
}
\value{
A data frame containing number of unique recombination events at genomic positions where recombination events were identified
}
\description{
This function reads a GFF file or data frame containing identified
recombination events in the genome identified by Gubbins, and counts the frequency of recombination
events at each genomic position i.e. number of unique overlapping recombination events. The
data frame can be generated using the "load.gubbins.rec.events.gff" function
}
\examples{
\dontrun{
Read genome in GFF formatted file (generated usign readseq) and plot
the genomic features

This function may take some minutes to finish depending on the number
of recombination events identifiedand genome size

gubbins.gff<-system.file("extdata", "ST320.recombination_predictions.gff",
package = "RCandy",mustWork = TRUE)

rec.freq<-count.rec.events.per.base(gubbins.gff)
}

}
\references{
\url{https://github.com/ChrispinChaguza/RCandy}
}
\author{
Chrispin Chaguza, \email{Chrispin.Chaguza@gmail.com}
}
