% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.colour.R
\name{is.color}
\alias{is.color}
\title{Check if string is a valid colour name}
\usage{
is.color(colour.name)
}
\arguments{
\item{colour.name}{Input character or vector containing colour names}
}
\value{
A Boolean showing whether or not a string or vector contains valid colour names
}
\description{
This function checks if a string or a vector contains valid colour names
}
\examples{
\dontrun{
Read the colour names as a string or vector of characters

col.names<-is.color(c("red","blue"))
col.names<-is.color("blue")
}

}
