% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.ref.genome.GFF.R
\name{load.genome.GFF}
\alias{load.genome.GFF}
\title{This function loads a reference genome file in GFF format.}
\usage{
load.genome.GFF(reference.genome)
}
\arguments{
\item{reference.genome}{Input file name or data frame in GFF file format.}
}
\value{
A data frame.
}
\description{
This function loads a reference genome file in GFF format.
}
\examples{
\dontrun{
Read genome in GFF formatted file, generated usign readseq, and plot the genomic features

ref.genome.gff<-system.file("extdata", "Hungary19A-6.gff", package = "RCandy",mustWork = TRUE)

new.ref.genome<-load.genome.GFF(ref.genome.gff)
}

}
\references{
\url{https://github.com/ChrispinChaguza/RCandy}
}
\author{
Chrispin Chaguza, \email{Chrispin.Chaguza@gmail.com}
}
