% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.gubbins.GFF.R
\name{load.gubbins.GFF}
\alias{load.gubbins.GFF}
\title{Read recombination events from the output GFF file from Gubbins}
\usage{
load.gubbins.GFF(gubbins.gff.file, recom.input.type = "Gubbins")
}
\arguments{
\item{gubbins.gff.file}{Path to the input Gubbins GFF recombination file or data frame}

\item{recom.input.type}{Type of input recombination data, either "Gubbins" GFF or "BRATNextGen" tabular data.}
}
\value{
A data frame containing number of unique recombination events at genomic positions where recombination events were identified
}
\description{
This function reads a GFF format generated by Gubbins containing
genomic regions where recombination events were identified.
}
\examples{
\dontrun{
Read genome in GFF formatted file (generated usign readseq) and plot the genomic features

gubbins.gff<-system.file("extdata", "ST320.recombination_predictions.gff", package = "RCandy",
mustWork = TRUE)

rec.events<-load.gubbins.rec.events.gff(gubbins.gff)
}

}
\references{
\url{https://github.com/ChrispinChaguza/RCandy}
}
\author{
Chrispin Chaguza, \email{Chrispin.Chaguza@gmail.com}
}
