% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.taxon.metadata.R
\name{load.taxon.metadata}
\alias{load.taxon.metadata}
\title{Function to read and process taxon metadata file}
\usage{
load.taxon.metadata(
  taxon.metadata.file,
  taxon.metadata.columns = NULL,
  taxon.names = NULL,
  taxon.id.column = NULL,
  include.first.col = FALSE,
  taxon.metadata.delimeter = "\\t"
)
}
\arguments{
\item{taxon.metadata.file}{Path to the input metadata file name or data frame.}

\item{taxon.metadata.columns}{A vector containing name of columns in the matadata file or data frame to view in the phylogenetic tree.}

\item{taxon.names}{A vector containing taxon names to select from the metadata file or data frames. These names must match the taxon names in the phylogenetic tree.}

\item{taxon.id.column}{Column name in the matadata file or data frame containing taxon names.}

\item{include.first.col}{A Boolean value specifying whether to use the first column as the taxon names.}

\item{taxon.metadata.delimeter}{A delimiter separating metadata columns.}
}
\value{
A data frame containing selected metadata columns and taxon names in the phylogenetic tree.
}
\description{
This function reads and process a user specified metadata file or data frame.
It assumes that the file is in text format and that the columns are tab-delimited.
Metadata provided in any other format other than as a 'character' class file name or
a data frame with class "tbl_df", "tbl","grouped_df","data.frame" or "rowwise_df" will not be
accepted. When not specified, it will assume that the first column represents
the taxon names. The taxon names should match those included in the
phylogenetic tree file or object
}
\examples{
\dontrun{
Read a tab-delimited file containing metadata

metadata.file<-system.file("extdata", "ST320.tsv", package = "RCandy",mustWork = TRUE)

metadata.df<-load.taxon.metadata(metadata.file)
}
}
\references{
\url{https://github.com/ChrispinChaguza/RCandy}
}
\author{
Chrispin Chaguza, \email{Chrispin.Chaguza@gmail.com}
}
