% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.phylogenetic.tree.R
\name{read.tree.file}
\alias{read.tree.file}
\title{Function to read phylogenetic tree in Newick file}
\usage{
read.tree.file(tree.file.name)
}
\arguments{
\item{tree.file.name}{Path to the input phylogenetic tree file in Newick format.}
}
\value{
A data frame containing selected metadata columns and strain names in the phylogenetic tree.
}
\description{
This function reads and process a user specified phylogenetic tree
file in Newick format.
}
\examples{
\dontrun{
Load phylogenetic tree file.

tree.file<-system.file("extdata", "ST320.final_tree.tre", package = "RCandy",mustWork = TRUE)

read.tree.file(tree.file.name=tree.file)
}

}
\references{
\url{https://github.com/ChrispinChaguza/RCandy}
}
\author{
Chrispin Chaguza, \email{Chrispin.Chaguza@gmail.com}
}
