% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.plot.R
\name{show.blank.plot}
\alias{show.blank.plot}
\title{Generate a blank plot}
\usage{
show.blank.plot()
}
\description{
This function generates a blank plot. The plot can be used as a filler in
when creating multipanel diagrams.
}
