% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.genome.annotations.R
\name{show.genome.annotation.plot}
\alias{show.genome.annotation.plot}
\title{Draw genome annotation features}
\usage{
show.genome.annotation.plot(
  genome.name,
  genome.start = NULL,
  genome.end = NULL,
  genome.start.upstream = 0,
  genome.end.downstream = 0,
  show.gene.label = FALSE,
  gene.feature.width = 1.5
)
}
\arguments{
\item{genome.name}{Path to the input file name.}

\item{genome.start}{Start position of the genome to show in the plot.}

\item{genome.end}{End position of the genome to show in the plot.}

\item{genome.start.upstream}{Start drawing the genome plot from the specified bases upstream of the genome.}

\item{genome.end.downstream}{End drawing the genome plot from the specified bases upstream of the genome.}

\item{show.gene.label}{Show genetic feature label.}

\item{gene.feature.width}{Width of the genetic features.}
}
\value{
None
}
\description{
This function reads a reference genome in GFF format and then plots the
genetic features (coding sequences) on both forward and reverse strands.
}
\examples{
\dontrun{
Read genome in GFF formatted file (generated usign readseq) and plot the genomic features

ref.genome.gff<-system.file("extdata", "Hungary19A-6.gff", package = "RCandy",mustWork = TRUE)

show.genome.annotation.plot(ref.genome.gff)
}

}
\references{
\url{https://github.com/ChrispinChaguza/RCandy}
}
\author{
Chrispin Chaguza, \email{Chrispin.Chaguza@gmail.com}
}
